refresh IaC scripts with latest conventions

+++ tools/infrastructure/ps-functions/Remove-AksApp.ps1
@@ -20,7 +20,9 @@ function Remove-AksApp {
         [Parameter(Mandatory)]
         [string] $Namespace,
     
-        [string[]] $ManagedIdentityName = @()
+        [string[]] $ManagedIdentityName = @(),
+    
+        [switch] $PodIdentityOnly
     )
     begin {
         Set-StrictMode -Version 'Latest'
@@ -35,6 +37,11 @@ function Remove-AksApp {
                 ResourceGroupName   =   $AksResourceGroup
             }
         }
+
+        Write-Information "Configuring auto-install of az-preview extension..."
+        Invoke-Exe {
+            az config set extension.use_dynamic_install=yes_without_prompt
+        } | Out-Null
     }
     process {
         try {            
@@ -45,32 +52,18 @@ function Remove-AksApp {
                 az aks get-credentials -g $aks.ResourceGroupName -n $aks.ResourceName --overwrite-existing
             } | Out-Null
 
-            Write-Information "Getting details of Node Resource Group of AKS cluster '$($aks.ResourceName)'..."
-            $nodeRg = Invoke-Exe {
-                az aks show -g $aks.ResourceGroupName -n $aks.ResourceName --query nodeResourceGroup -o tsv
+            if (-not($PodIdentityOnly)) {
+                Write-Information "Uninstalling helm chart release '$HelmChartName' in namespace '$Namespace'..."
+                Invoke-Exe { helm uninstall $HelmChartName --namespace $Namespace }  -EA Continue    
             }
-            $nodeRgResourceId = Invoke-Exe { az group show -n $nodeRg -o tsv --query 'id' }
 
-            Write-Information "Uninstalling helm chart release '$HelmChartName' in namespace '$Namespace'..."
-            Invoke-Exe { helm uninstall $HelmChartName --namespace $Namespace }  -EA Continue
-
-            $managedIdentityName | ForEach-Object {
+            $managedIdentityName | Where-Object { $_ } | ForEach-Object {
                 $name = $_
                 
                 Write-Information "Deleting Pod Identity '$name' from AKS cluster '$($aks.ResourceName)'..."
                 Invoke-Exe {
                     az aks pod-identity delete --name $name --namespace $Namespace -g $aks.ResourceGroupName --cluster-name $aks.ResourceName
                 }  -EA Continue | Out-Null
-                
-                $identityClientId = Invoke-Exe {
-                    az identity show -g $AppResourceGroup -n $name --query clientId -otsv
-                }
-                
-                $rbacRole = 'Virtual Machine Contributor'
-                Write-Information "Removing  RBAC role '$rbacRole' from Managed Identity '$name' for the cluster node resource group..."
-                Invoke-Exe {
-                    az role assignment delete --assignee $identityClientId --role $rbacRole --scope $nodeRgResourceId
-                } -EA Continue | Out-Null
             }
         }
         catch
